requirements.txt
langchain
langchain-openai
langgraph
streamlit
pandas
plotly
python-dotenv
duckduckgo-search